feat(web): implement render callback as props to reuse task-badge
