--- web/src/app/dashboard/workspace/[workspace]/@modal/(.)task/[task]/_components/task-add-label/types.ts
@@ -1,27 +1,0 @@
-export type ColorTone = "bold" | "default" | "subtle";
-
-export const colorOrder = [
-  "slate",
-  "stone",
-  "zinc",
-  "red",
-  "orange",
-  "amber",
-  "lime",
-  "green",
-  "teal",
-  "cyan",
-  "blue",
-  "purple",
-] as const;
-
-export type ColorOrder = (typeof colorOrder)[number];
-
-export type ColorOption = {
-  name: ColorOrder;
-  value: {
-    bold: string;
-    default: string;
-    subtle: string;
-  };
-};
